fix: expose actions and errors as subpackages
