+++ src/errors/index.ts
@@ -1,0 +1,74 @@
+import {
+  AccountConfigurationNotSupportedError,
+  AccountError,
+  Eip7702AccountMustHaveEoaError,
+  Eip7702NotSupportedForAccountError,
+  ExistingEip7702AccountsNotSupportedError,
+  FactoryArgsNotAvailableError,
+  isAccountError,
+  SigningNotSupportedForAccountError,
+  SmartSessionsNotEnabledError,
+  WalletClientNoConnectedAccountError,
+} from '../accounts'
+import {
+  ExecutionError,
+  IntentFailedError,
+  isExecutionError,
+  OrderPathRequiredForIntentsError,
+  SessionChainRequiredError,
+  SimulationNotSupportedForUserOpFlowError,
+  SourceChainsNotAvailableForUserOpFlowError,
+  UserOperationRequiredForSmartSessionsError,
+} from '../execution'
+import {
+  AuthenticationRequiredError,
+  InsufficientBalanceError,
+  IntentNotFoundError,
+  InvalidApiKeyError,
+  InvalidIntentSignatureError,
+  isOrchestratorError,
+  NoPathFoundError,
+  OnlyOneTargetTokenAmountCanBeUnsetError,
+  OrchestratorError,
+  TokenNotSupportedError,
+  UnsupportedChainError,
+  UnsupportedChainIdError,
+  UnsupportedTokenError,
+} from '../orchestrator'
+
+export {
+  // Account
+  isAccountError,
+  AccountError,
+  AccountConfigurationNotSupportedError,
+  Eip7702AccountMustHaveEoaError,
+  ExistingEip7702AccountsNotSupportedError,
+  FactoryArgsNotAvailableError,
+  SmartSessionsNotEnabledError,
+  SigningNotSupportedForAccountError,
+  Eip7702NotSupportedForAccountError,
+  WalletClientNoConnectedAccountError,
+  // Execution
+  isExecutionError,
+  ExecutionError,
+  IntentFailedError,
+  OrderPathRequiredForIntentsError,
+  SessionChainRequiredError,
+  SimulationNotSupportedForUserOpFlowError,
+  SourceChainsNotAvailableForUserOpFlowError,
+  UserOperationRequiredForSmartSessionsError,
+  // Orchestrator
+  isOrchestratorError,
+  AuthenticationRequiredError,
+  InsufficientBalanceError,
+  InvalidApiKeyError,
+  InvalidIntentSignatureError,
+  NoPathFoundError,
+  OnlyOneTargetTokenAmountCanBeUnsetError,
+  OrchestratorError,
+  IntentNotFoundError,
+  TokenNotSupportedError,
+  UnsupportedChainError,
+  UnsupportedChainIdError,
+  UnsupportedTokenError,
+}
