+++ src/index.ts
@@ -9,73 +9,22 @@ import type {
 } from 'viem'
 import type { UserOperationReceipt } from 'viem/account-abstraction'
 import {
-  AccountConfigurationNotSupportedError,
-  AccountError,
   checkAddress,
   deploy as deployInternal,
-  Eip7702AccountMustHaveEoaError,
-  Eip7702NotSupportedForAccountError,
-  ExistingEip7702AccountsNotSupportedError,
-  FactoryArgsNotAvailableError,
   getAddress as getAddressInternal,
-  isAccountError,
   isDeployed as isDeployedInternal,
-  SigningNotSupportedForAccountError,
-  SmartSessionsNotEnabledError,
   setup as setupInternal,
   signEip7702InitData as signEip7702InitDataInternal,
-  WalletClientNoConnectedAccountError,
 } from './accounts'
 import { walletClientToAccount } from './accounts/walletClient'
+import { encodeSmartSessionSignature } from './actions/smart-session'
 import {
-  addOwner,
-  addPasskeyOwner,
-  changeMultiFactorThreshold,
-  changePasskeyThreshold,
-  changeThreshold,
-  disableEcdsa,
-  disableMultiFactor,
-  disablePasskeys,
-  enableEcdsa,
-  enableMultiFactor,
-  enablePasskeys,
-  encodeSmartSessionSignature,
-  installModule,
-  recover,
-  recoverEcdsaOwnership,
-  recoverPasskeyOwnership,
-  removeOwner,
-  removePasskeyOwner,
-  removeSubValidator,
-  setSubValidator,
-  setUpRecovery,
-  uninstallModule,
-} from './actions'
-import {
-  ExecutionError,
   getMaxSpendableAmount as getMaxSpendableAmountInternal,
   getPortfolio as getPortfolioInternal,
-  IntentFailedError,
-  isExecutionError,
-  OrderPathRequiredForIntentsError,
-  SessionChainRequiredError,
-  SimulationNotSupportedForUserOpFlowError,
-  SourceChainsNotAvailableForUserOpFlowError,
   sendTransaction as sendTransactionInternal,
   type TransactionResult,
-  UserOperationRequiredForSmartSessionsError,
   waitForExecution as waitForExecutionInternal,
 } from './execution'
-import {
-  depositErc20,
-  depositEther,
-  disableErc20Withdrawal,
-  disableEtherWithdrawal,
-  enableErc20Withdrawal,
-  enableEtherWithdrawal,
-  withdrawErc20,
-  withdrawEther,
-} from './execution/compact'
 import {
   getSessionDetails as getSessionDetailsInternal,
   type SessionDetails,
@@ -97,30 +46,17 @@ import {
   getValidators as getValidatorsInternal,
 } from './modules'
 import {
-  AuthenticationRequiredError,
   getSupportedTokens,
   getTokenAddress,
-  InsufficientBalanceError,
   type IntentCost,
   type IntentInput,
-  IntentNotFoundError,
   type IntentOp,
   type IntentOpStatus,
   type IntentResult,
   type IntentRoute,
-  InvalidApiKeyError,
-  InvalidIntentSignatureError,
-  isOrchestratorError,
-  NoPathFoundError,
-  OnlyOneTargetTokenAmountCanBeUnsetError,
-  OrchestratorError,
   type Portfolio,
   type SettlementSystem,
   type SignedIntentOp,
-  TokenNotSupportedError,
-  UnsupportedChainError,
-  UnsupportedChainIdError,
-  UnsupportedTokenError,
 } from './orchestrator'
 import type {
   AccountProviderConfig,
@@ -494,73 +430,8 @@ class RhinestoneSDK {
 
 export {
   RhinestoneSDK,
-  // Helpers
   walletClientToAccount,
-  // Actions
-  addOwner,
-  addPasskeyOwner,
-  changeMultiFactorThreshold,
-  changeThreshold,
-  changePasskeyThreshold,
-  disableEcdsa,
-  disableMultiFactor,
-  disablePasskeys,
-  enableEcdsa,
-  enableMultiFactor,
-  enablePasskeys,
   encodeSmartSessionSignature,
-  installModule,
-  recover,
-  recoverEcdsaOwnership,
-  recoverPasskeyOwnership,
-  removeOwner,
-  removePasskeyOwner,
-  removeSubValidator,
-  setSubValidator,
-  setUpRecovery,
-  uninstallModule,
-  depositErc20,
-  depositEther,
-  disableErc20Withdrawal,
-  disableEtherWithdrawal,
-  enableErc20Withdrawal,
-  enableEtherWithdrawal,
-  withdrawErc20,
-  withdrawEther,
-  // Account errors
-  isAccountError,
-  AccountError,
-  AccountConfigurationNotSupportedError,
-  Eip7702AccountMustHaveEoaError,
-  ExistingEip7702AccountsNotSupportedError,
-  FactoryArgsNotAvailableError,
-  SmartSessionsNotEnabledError,
-  SigningNotSupportedForAccountError,
-  Eip7702NotSupportedForAccountError,
-  WalletClientNoConnectedAccountError,
-  // Execution errors
-  isExecutionError,
-  ExecutionError,
-  IntentFailedError,
-  OrderPathRequiredForIntentsError,
-  SessionChainRequiredError,
-  SimulationNotSupportedForUserOpFlowError,
-  SourceChainsNotAvailableForUserOpFlowError,
-  UserOperationRequiredForSmartSessionsError,
-  // Orchestrator errors
-  isOrchestratorError,
-  AuthenticationRequiredError,
-  InsufficientBalanceError,
-  InvalidApiKeyError,
-  InvalidIntentSignatureError,
-  NoPathFoundError,
-  OnlyOneTargetTokenAmountCanBeUnsetError,
-  OrchestratorError,
-  IntentNotFoundError,
-  TokenNotSupportedError,
-  UnsupportedChainError,
-  UnsupportedChainIdError,
-  UnsupportedTokenError,
   // Registry functions
   getSupportedTokens,
   getTokenAddress,
